+++ vm/src/hint_processor/builtin_hint_processor/ec_utils.rs
@@ -1,4 +1,4 @@
-use crate::stdlib::{borrow::Cow, boxed::Box, collections::HashMap, prelude::*};
+use crate::stdlib::{boxed::Box, collections::HashMap, prelude::*};
 use crate::utils::{bigint_to_felt, biguint_to_felt, felt_to_biguint, CAIRO_PRIME};
 use crate::Felt252;
 use crate::{
@@ -20,17 +20,17 @@ use sha2::{Digest, Sha256};
 use super::hint_utils::get_ptr_from_var_name;
 
 #[derive(Debug, PartialEq)]
-struct EcPoint<'a> {
-    x: Cow<'a, Felt252>,
-    y: Cow<'a, Felt252>,
+struct EcPoint {
+    x: Felt252,
+    y: Felt252,
 }
-impl EcPoint<'_> {
-    fn from_var_name<'a>(
-        name: &'a str,
-        vm: &'a VirtualMachine,
-        ids_data: &'a HashMap<String, HintReference>,
-        ap_tracking: &'a ApTracking,
-    ) -> Result<EcPoint<'a>, HintError> {
+impl EcPoint {
+    fn from_var_name(
+        name: &str,
+        vm: &VirtualMachine,
+        ids_data: &HashMap<String, HintReference>,
+        ap_tracking: &ApTracking,
+    ) -> Result<EcPoint, HintError> {
         // Get first addr of EcPoint struct
         let point_addr = get_relocatable_from_var_name(name, vm, ids_data, ap_tracking)?;
         Ok(EcPoint {
@@ -109,7 +109,7 @@ pub fn chained_ec_op_random_ec_point_hint(
 ) -> Result<(), HintError> {
     let n_elms = get_integer_from_var_name("len", vm, ids_data, ap_tracking)?;
     if n_elms.is_zero() || n_elms.to_usize().is_none() {
-        return Err(HintError::InvalidLenValue(Box::new(n_elms.into_owned())));
+        return Err(HintError::InvalidLenValue(Box::new(n_elms)));
     }
     let n_elms = n_elms.to_usize().unwrap();
     let p = EcPoint::from_var_name("p", vm, ids_data, ap_tracking)?;
@@ -141,7 +141,7 @@ pub fn recover_y_hint(
     ids_data: &HashMap<String, HintReference>,
     ap_tracking: &ApTracking,
 ) -> Result<(), HintError> {
-    let p_x = get_integer_from_var_name("x", vm, ids_data, ap_tracking)?.into_owned();
+    let p_x = get_integer_from_var_name("x", vm, ids_data, ap_tracking)?;
     let p_addr = get_relocatable_from_var_name("p", vm, ids_data, ap_tracking)?;
     vm.insert_value(p_addr, p_x)?;
     let p_y = biguint_to_felt(